+++ installers/common/Types.hs
@@ -143,6 +143,9 @@ data InstallerConfig = InstallerConfig {
     , macPackageName :: Text
     , dataDir :: Text
     , hasBlock0 :: Bool
+    , genesisPath :: Maybe Text
+    , secretPath :: Maybe Text
+    , configPath :: Maybe Text
     } deriving (Generic, Show)
 
 instance FromJSON InstallerConfig where